ocr_nutrition_app/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
Pillow>=9.2  # For image processing
redis==5.2.0
django-redis==5.4.0
beautifulsoup4==4.12.3
bs4==0.0.2
pyzbar>=0.1.9